+++ CHANGELOG.md
@@ -31,6 +31,18 @@ This project uses [Semantic Versioning](https://semver.org/).
 
 ---
 
+## [0.7.0] - 2026-04-25
+
+### Added
+
+- Added IPC handlers for player storage, quest state storage, and quest definition loading.
+
+### Changed
+
+- Updated visible application version labels to `0.7.0`.
+
+---
+
 ## [0.6.0] - 2026-04-25
 
 ### Added
@@ -153,7 +165,8 @@ This project uses [Semantic Versioning](https://semver.org/).
 
 ---
 
-[Unreleased]: https://github.com/lotoos0/sol_hud/compare/v0.6.0...HEAD
+[Unreleased]: https://github.com/lotoos0/sol_hud/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/lotoos0/sol_hud/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/lotoos0/sol_hud/compare/v0.5.2...v0.6.0
 [0.5.2]: https://github.com/lotoos0/sol_hud/compare/v0.5.1...v0.5.2
 [0.5.1]: https://github.com/lotoos0/sol_hud/compare/v0.5.0...v0.5.1